Add size check, remove poly1305 helper
