+++ src/lib.rs
@@ -145,35 +145,6 @@ pub fn chacha20_block(key: &Secret<[u8; 32]>, counter: u32, nonce: &[u8; 12]) ->
     chacha20_block_bytes(key.expose_secret(), counter, nonce)
 }
 
-pub fn poly1305_tag(r: &u128, s: &u128, aad: &[u8], ciphertext: &[u8]) -> [u8; 16] {
-    use poly1305::{
-        universal_hash::{KeyInit, UniversalHash},
-        Block, Key, Poly1305,
-    };
-
-    let mut key_bytes = [0u8; 32];
-    key_bytes[..16].copy_from_slice(&r.to_le_bytes());
-    key_bytes[16..].copy_from_slice(&s.to_le_bytes());
-    lock(&key_bytes).ok();
-    let mut poly = Poly1305::new(Key::from_slice(&key_bytes));
-
-    poly.update_padded(aad);
-    poly.update_padded(ciphertext);
-
-    let mut len_block = [0u8; 16];
-    len_block[..8].copy_from_slice(&(aad.len() as u64).to_le_bytes());
-    len_block[8..].copy_from_slice(&(ciphertext.len() as u64).to_le_bytes());
-    poly.update(&[Block::clone_from_slice(&len_block)]);
-
-    let tag = poly.finalize();
-    let mut out = [0u8; 16];
-    out.copy_from_slice(tag.as_slice());
-    unlock(&key_bytes).ok();
-    key_bytes.zeroize();
-    len_block.zeroize();
-    out
-}
-
 use subtle::ConstantTimeEq;
 
 pub fn ct_eq(a: &[u8], b: &[u8]) -> bool {
@@ -181,6 +152,10 @@ pub fn ct_eq(a: &[u8], b: &[u8]) -> bool {
 }
 
 /// Read an entire file while using the same code path on success or failure.
+///
+/// This aims to keep timing similar for error and success cases but it is not a
+/// strong constant-time guarantee. True constant-time I/O would require
+/// operating system support and is outside the scope of this crate.
 pub fn read_file_ct(path: &Path) -> Result<Vec<u8>> {
     let mut buf = Vec::new();
     let dummy = [0u8; 1];
