+++ tests/encryption.rs
@@ -1,7 +1,26 @@
-use encryptor::{
-    chacha20_block, ct_eq, derive_key, encrypt_decrypt, poly1305_tag, read_file_ct, Argon2Config,
+use encryptor::{chacha20_block, ct_eq, derive_key, encrypt_decrypt, read_file_ct, Argon2Config};
+use poly1305::{
+    universal_hash::{KeyInit, UniversalHash},
+    Block, Key, Poly1305,
 };
 use secrecy::ExposeSecret;
+
+fn compute_tag(r: &u128, s: &u128, aad: &[u8], ciphertext: &[u8]) -> [u8; 16] {
+    let mut key_bytes = [0u8; 32];
+    key_bytes[..16].copy_from_slice(&r.to_le_bytes());
+    key_bytes[16..].copy_from_slice(&s.to_le_bytes());
+    let mut poly = Poly1305::new(Key::from_slice(&key_bytes));
+    poly.update_padded(aad);
+    poly.update_padded(ciphertext);
+    let mut len_block = [0u8; 16];
+    len_block[..8].copy_from_slice(&(aad.len() as u64).to_le_bytes());
+    len_block[8..].copy_from_slice(&(ciphertext.len() as u64).to_le_bytes());
+    poly.update(&[Block::clone_from_slice(&len_block)]);
+    let tag = poly.finalize();
+    let mut out = [0u8; 16];
+    out.copy_from_slice(tag.as_slice());
+    out
+}
 
 #[test]
 fn encrypt_decrypt_roundtrip() {
@@ -40,10 +59,10 @@ fn poly1305_tag_detects_modification() {
     let header = b"header";
     let plaintext = b"message";
     let mut ciphertext = encrypt_decrypt(plaintext, &key, &nonce);
-    let tag = poly1305_tag(&r, &s, header, &ciphertext);
+    let tag = compute_tag(&r, &s, header, &ciphertext);
     // flip a byte in ciphertext
     ciphertext[0] ^= 1;
-    let wrong_tag = poly1305_tag(&r, &s, header, &ciphertext);
+    let wrong_tag = compute_tag(&r, &s, header, &ciphertext);
     assert!(!ct_eq(&tag, &wrong_tag));
 }
 
@@ -71,11 +90,11 @@ fn tampered_ciphertext_fails_to_authenticate() {
     let header = b"hdr";
     let plaintext = b"secret";
     let ciphertext = encrypt_decrypt(plaintext, &key, &nonce);
-    let tag = poly1305_tag(&r, &s, header, &ciphertext);
+    let tag = compute_tag(&r, &s, header, &ciphertext);
 
     let mut tampered = ciphertext.clone();
     tampered[1] ^= 0x80;
-    let expected = poly1305_tag(&r, &s, header, &tampered);
+    let expected = compute_tag(&r, &s, header, &tampered);
     assert!(!ct_eq(&tag, &expected));
     let decrypted = encrypt_decrypt(&tampered, &key, &nonce);
     assert_ne!(plaintext.to_vec(), decrypted);
@@ -105,9 +124,9 @@ fn tampered_tag_fails_to_authenticate() {
     let header = b"hdr2";
     let plaintext = b"data";
     let ciphertext = encrypt_decrypt(plaintext, &key, &nonce);
-    let mut tag = poly1305_tag(&r, &s, header, &ciphertext);
+    let mut tag = compute_tag(&r, &s, header, &ciphertext);
     tag[0] ^= 0x01;
-    let expected = poly1305_tag(&r, &s, header, &ciphertext);
+    let expected = compute_tag(&r, &s, header, &ciphertext);
     assert!(!ct_eq(&tag, &expected));
 }
